code/requirements.txt
# Microsoft Agent Framework (MAF) with Azure AI Foundry integration
agent-framework-foundry>=1.0.1

# Azure AI Foundry Projects SDK (for agent management scripts)
azure-ai-projects>=2.0.0

# Azure Identity
azure-identity>=1.15.0

# Data models
pydantic>=2.0.0

# Rich console UI (Spectre-style progress)
rich>=13.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
